Remove @RepeatableSuppressWarnings stuff
