+++ gradle-suppressible-error-prone/build.gradle
@@ -39,8 +39,6 @@ test {
 
 // Added as a jar the errorprone configuration inside the plugin, so in nebula-tests needs to exist in maven local
 tasks.test.dependsOn tasks.findByPath(':suppressible-error-prone:publishToMavenLocal')
-// Similarly, added as a jar for compileOnly inside the plugin, so in nebula-tests needs to exist in maven local
-tasks.test.dependsOn tasks.findByPath(':suppressible-error-prone-annotations:publishToMavenLocal')
 
 tasks.withType(JavaCompile) {
     options.errorprone {
