+++ settings.gradle
@@ -13,4 +13,3 @@ rootProject.name = "suppressible-error-prone-root"
 
 include 'gradle-suppressible-error-prone'
 include 'suppressible-error-prone'
-include 'suppressible-error-prone-annotations'
