--- suppressible-error-prone-annotations/src/main/java/com/palantir/suppressibleerrorprone/RepeatableSuppressWarningsContainer.java
@@ -1,33 +1,0 @@
-/*
- * (c) Copyright 2024 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.suppressibleerrorprone;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Target;
-
-@Target({
-    ElementType.TYPE,
-    ElementType.FIELD,
-    ElementType.METHOD,
-    ElementType.PARAMETER,
-    ElementType.CONSTRUCTOR,
-    ElementType.LOCAL_VARIABLE,
-    ElementType.MODULE
-})
-public @interface RepeatableSuppressWarningsContainer {
-    RepeatableSuppressWarnings[] value();
-}
