+++ suppressible-error-prone/build.gradle
@@ -11,7 +11,6 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter'
     testImplementation 'com.google.errorprone:error_prone_test_helpers'
     testImplementation 'org.assertj:assertj-core'
-    testRuntimeOnly project(':suppressible-error-prone-annotations')
 }
 
 moduleJvmArgs {
